przygotowanie do finalnego drona
Przeszkody ani drony nie są porozmieszczane, brak możliwości zmiany drona

+++ inc/Plaszczyzna.h
@@ -8,9 +8,9 @@ using drawNS::Point3D;
 class Plaszczyzna:public ObiektRysowalny{
 
 protected:
-    float wspolrzednaZ;
-    float wielkoscPlaszczyzny;
-    float wielkoscKratki;
+    double wspolrzednaZ;
+    double wielkoscPlaszczyzny;
+    double wielkoscKratki;
